+++ nym-vpn-core/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Windows] App Split Tunnelling (https://github.com/nymtech/nym-vpn-client/pull/4908).
 - CLI: add command to list processes excluded from VPN tunnel: `nym-vpnc split-tunnel excluded-processes`
+- Connection middleware layer (https://github.com/nymtech/nym-vpn-client/pull/5000)
 
 ### Fixed